Add Parallel::ForkManager::DummyProcessStatus
This lets us hide the bit shifting & masking used to decode a process's
exit status to see if the process exited due to a signal or if it dumped
a core.

This was added because it's not obvious how to create a Process::Status
object.  The DummyProcessStatus only provides the bits we're interested
in.

"Incidental" changes:

Make sure that the serializer is always set up, even if the constructor
params hash is empty or not provided

Add some error classes so the error classname is descriptive and we have
a set of Parallel::ForkManager::Error children.

Make in_child a private attr_accessor.